adicionando mensagem amigável para o usuário

+++ Program.cs
@@ -1,74 +1,105 @@
 ﻿using Microsoft.Data.Sqlite;
-using LabManager.DataBase;
-using LabManager.Repositories; //importar a classe 
+using LabManager.DataBase; //importar classes com using 
+using LabManager.Repositories;
 using LabManager.Models;
 
-var databaseConfig = new DatabaseConfig(); //objeto. não precisa ficar criando nas classes
+//program.cs não precisa de namespace pq eh o principal
 
-var databaseSetup = new DatabaseSetup(databaseConfig);
-
-var computerRepository = new ComputerRepository(databaseConfig); //criar a váriavel 
+var databaseConfig = new DatabaseConfig(); //criando os objetos database
+var databaseSetup = new DatabaseSetup(databaseConfig); 
+var computerRepository = new ComputerRepository(databaseConfig);
 var labRepository = new LabRepository(databaseConfig);
 
-//Routing
-var modelName = args[0];
+var modelName = args[0]; //armazena se o comando é computer ou lab
 var modelAction = args[1];
 
 if (modelName == "Computer")
 {
-        if (modelAction == "List")
+    if (modelAction == "List")
+    {
+        Console.WriteLine("Computer List");
+        foreach (var computer in computerRepository.GetAll())
         {
-                Console.WriteLine("Computer List");
-                foreach (var computer in computerRepository.GetAll())
-                {
-                        Console.WriteLine(
-                                "{0}, {1}, {2}", computer.Id, computer.Ram, computer.Processor
-                                );
-                }
+            Console.WriteLine(
+                "{0}, {1}, {2}", computer.Id, computer.Ram, computer.Processor
+            );
+        }
+    }
+
+    if (modelAction == "New")
+    {
+        var id = Convert.ToInt32(args [2]);
+        var ram = (args [3]);
+        var processor = (args [4]);
+        var computer = new Computer (id, ram, processor);
+
+        if (computerRepository.ExistsById(id))
+        {
+            Console.WriteLine ("O computador com o ID {0} já existe.", id);
+            
+        }
+        else 
+        {
+            computerRepository.Save(computer);
+        } 
+    }
+
+    if (modelAction == "Delete")
+    {
+        var id = Convert.ToInt32 (args [2]);
+
+
+        if (computerRepository.ExistsById(id))
+        {
+            computerRepository.Delete (id);
+        }
+        else 
+        {
+            Console.WriteLine ("O computador com o ID {0} não existe.", id);
         }
 
-        if (modelAction == "New") 
-        {
-                var id = Convert.ToInt32(args [2]);
-                var ram = args [3];
-                var processor = args [4];
+    }
 
-                var computer = new Computer(id, ram, processor);
-                computerRepository.Save(computer);
-        }
-        
-        if (modelAction == "Show") 
-        {
-                var id = Convert.ToInt32(args [2]);
-                var computer = computerRepository.GetById(id);
-                Console.WriteLine("{0}, {1}, {2}", computer.Id, computer.Ram, computer.Processor);
-        }
+    if (modelAction == "Show")
+    {
+        var id = Convert.ToInt32 (args [2]);
 
-        if (modelAction == "Delete") 
+        if (computerRepository.ExistsById(id))
         {
-                var id = Convert.ToInt32(args [2]);
-                computerRepository.Delete(id);
+            var computer = computerRepository.GetById(id);
+            Console.WriteLine ("{0}, {1}, {2}", computer.Id, computer.Ram, computer.Processor);
         }
-
-        if (modelAction == "Update") 
+        else 
         {
-                var id = Convert.ToInt32(args [2]);
-                var ram = args [3];
-                var processor = args [4];
-
-                var computer = new Computer(id, ram, processor);
-                computerRepository.Update(computer);
-                Console.WriteLine("{0}, {1}, {2}", computer.Id, computer.Ram, computer.Processor);
+            Console.WriteLine ("O computador com o ID {0} não existe.", id);
         }
+    }
+
+    if (modelAction == "Update")
+    {
+        var id = Convert.ToInt32 (args [2]);
+        string ram = (args [3]);
+        string processor = (args [4]);
+        var computer =  new Computer (id, ram, processor);
+
+        if (computerRepository.ExistsById(id))
+        {
+            computerRepository.Update(computer);
+        }
+        else 
+        {
+            Console.WriteLine ("O computador com o ID {0} não existe.", id);
+        }
+    }
 }
 
-if(modelName == "Lab")
+if (modelAction == "Lab")
 {
-    if(modelAction == "List")
+    if (modelName == "List")
     {
         foreach (var lab in labRepository.GetAll())
         {
-            Console.WriteLine("{0}, {1}, {2}, {3}", lab.Id, lab.Number, lab.Name, lab.Block);
+            Console.WriteLine ("{0}, {1}, {2}, {3}", lab.Id, lab.Number, lab.Name, lab.Block);
         }
     }
 
@@ -78,18 +109,32 @@ if(modelName == "Lab")
         var number = Convert.ToInt32(args[3]);
         var name = args[4];
         var block = args[5];
-        Console.WriteLine("New Lab");
-        Console.WriteLine("{0}, {1}, {2}, {3}", id, number, name, block);
 
         var lab = new Lab(id, number, name, block);
-        labRepository.Save(lab);
+
+        if (labRepository.ExistsById(id))
+        {
+            Console.WriteLine ("O laboratório com o ID {0} já existe.", id); 
+        }
+        else 
+        {
+            labRepository.Save(lab);
+        } 
     }
 
     if (modelAction == "Show")
     {
         var id = Convert.ToInt32(args[2]);
-        var lab = labRepository.GetById(id);
-        Console.WriteLine("{0}, {1}, {2}, {3}", lab.Id, lab.Number, lab.Name, lab.Block);
+
+        if (labRepository.ExistsById(id))
+        {
+            var lab = labRepository.GetById(id);
+            Console.WriteLine("{0}, {1}, {2}, {3}", lab.Id, lab.Number, lab.Name, lab.Block);
+        }
+        else 
+        {
+            Console.WriteLine ("O laboratório com o ID {0} não existe.", id);
+        } 
     }
 
     if (modelAction == "Update")
@@ -98,14 +143,30 @@ if(modelName == "Lab")
         var number = Convert.ToInt32(args[3]);
         var name = args[4];
         var block = args[5];
-        var lab = new Lab(id, number, name, block);
+        
+        if (labRepository.ExistsById(id))
+        {
+            var lab = new Lab(id, number, name, block);
 
-        labRepository.Update(lab);
+            labRepository.Update(lab);
+        }
+        else 
+        {
+            Console.WriteLine ("O laboratório com o ID {0} não existe.", id);
+        } 
     }
 
     if (modelAction == "Delete")
     {
-        var id = Convert.ToInt32(args[2]);
-        labRepository.Delete(id);
+        var id = Convert.ToInt32(args[2]);  
+
+        if (labRepository.ExistsById(id))
+        {  
+            labRepository.Delete(id);
+        }
+        else 
+        {
+            Console.WriteLine ("O laboratório com o ID {0} não existe.", id);
+        } 
     }
 }
